Remove Cargo.lock
Due to differences in supplied library versions (libssh, libssl, etc.)
between my development machine and the deployment server this must
be rebuilt on the server for every update. The lockfile was causing
problems with the rebuild and had to be deleted and regenerated.
So the easiest solution is to just remove and ignore the file.
This may cause problems in the future with different dependency
versions but it works for now.